Move all SYCL headers into SYCL directory
