+++ core/src/SYCL/Kokkos_SYCL_Space.cpp
@@ -21,8 +21,8 @@
 #include <Kokkos_Macros.hpp>
 
 #include <Kokkos_HostSpace.hpp>
-#include <Kokkos_SYCL.hpp>
-#include <Kokkos_SYCL_Space.hpp>
+#include <SYCL/Kokkos_SYCL.hpp>
+#include <SYCL/Kokkos_SYCL_Space.hpp>
 #include <SYCL/Kokkos_SYCL_DeepCopy.hpp>
 #include <SYCL/Kokkos_SYCL_Instance.hpp>
 #include <impl/Kokkos_MemorySpace.hpp>
